Fixed corpse respawn when client sends empty guid. Fixed cooldown reduction from haste calculations. Fixed npc health calculations. Changed npc damage calculations. Added several more portal effects. Implemented return to graveyard functionality. Re-implemented player play time functionality.

+++ src/ronin-world/Mainframe/ObjectMgr.h
@@ -154,7 +154,7 @@ public:
     typedef std::map<uint32, SkillLineAbilityEntry*>             SLMap;
     typedef std::map<uint32, std::vector<VendorItem> >           VendorMap;
     typedef std::map<uint32, TrainerData >                       TrainerDataMap;
-    typedef std::map<std::pair<uint8, uint8>, TrainerSpellMap>   TrainerSpellStorage;
+    typedef std::map<std::pair<uint8, uint16>, TrainerSpellMap>  TrainerSpellStorage;
     typedef std::map<uint32, Corpse* >                           CorpseMap;
     typedef std::map<uint32, Group*>                             GroupMap;
 
@@ -263,7 +263,7 @@ public:
 
     // Trainers
     TrainerData *GetTrainerData(uint32 entry);
-    TrainerSpellMap *GetTrainerSpells(uint8 category, uint8 subcategory);
+    TrainerSpellMap *GetTrainerSpells(uint8 category, uint16 subcategory);
 
     uint32 GenerateArenaTeamId()
     {
